Improve glean tests by using `resetGlean` whenever possible
