+++ components/service/glean/src/test/java/mozilla/components/service/glean/DatetimeMetricTypeTest.kt
@@ -4,11 +4,8 @@
 
 package mozilla.components.service.glean
 
-import android.content.Context
-import androidx.test.core.app.ApplicationProvider
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.ObsoleteCoroutinesApi
-import mozilla.components.service.glean.storages.DatetimesStorageEngine
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
@@ -30,15 +27,7 @@ class DatetimeMetricTypeTest {
 
     @Before
     fun setUp() {
-        Glean.initialized = true
-        DatetimesStorageEngine.applicationContext = ApplicationProvider.getApplicationContext()
-        // Clear the stored "user" preferences between tests.
-        ApplicationProvider.getApplicationContext<Context>()
-            .getSharedPreferences(DatetimesStorageEngine.javaClass.canonicalName, Context.MODE_PRIVATE)
-            .edit()
-            .clear()
-            .apply()
-        DatetimesStorageEngine.clearAllStores()
+        resetGlean()
     }
 
     @Test
