+++ components/service/glean/src/test/java/mozilla/components/service/glean/StringListMetricTypeTest.kt
@@ -4,12 +4,9 @@
 
 package mozilla.components.service.glean
 
-import android.content.Context
-import androidx.test.core.app.ApplicationProvider
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.ObsoleteCoroutinesApi
 import mozilla.components.service.glean.StringListMetricType.Companion.MAX_STRING_LENGTH
-import mozilla.components.service.glean.storages.StringListsStorageEngine
 import mozilla.components.service.glean.storages.StringListsStorageEngineImplementation
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
@@ -31,17 +28,7 @@ class StringListMetricTypeTest {
 
     @Before
     fun setUp() {
-        Glean.initialized = true
-        StringListsStorageEngine.applicationContext = ApplicationProvider.getApplicationContext()
-        // Clear the stored "user" preferences between tests.
-        ApplicationProvider.getApplicationContext<Context>()
-            .getSharedPreferences(
-                StringListsStorageEngine.javaClass.canonicalName,
-                Context.MODE_PRIVATE)
-            .edit()
-            .clear()
-            .apply()
-        StringListsStorageEngine.clearAllStores()
+        resetGlean()
     }
 
     @Test
