+++ components/service/glean/src/test/java/mozilla/components/service/glean/storages/BooleansStorageEngineTest.kt
@@ -24,12 +24,6 @@ class BooleansStorageEngineTest {
     @Before
     fun setUp() {
         BooleansStorageEngine.applicationContext = ApplicationProvider.getApplicationContext()
-        // Clear the stored "user" preferences between tests.
-        ApplicationProvider.getApplicationContext<Context>()
-            .getSharedPreferences(BooleansStorageEngine.javaClass.canonicalName, Context.MODE_PRIVATE)
-            .edit()
-            .clear()
-            .apply()
         BooleansStorageEngine.clearAllStores()
     }
