+++ components/service/glean/src/test/java/mozilla/components/service/glean/storages/CountersStorageEngineTest.kt
@@ -25,12 +25,6 @@ class CountersStorageEngineTest {
     @Before
     fun setUp() {
         CountersStorageEngine.applicationContext = ApplicationProvider.getApplicationContext()
-        // Clear the stored "user" preferences between tests.
-        ApplicationProvider.getApplicationContext<Context>()
-            .getSharedPreferences(CountersStorageEngine.javaClass.canonicalName, Context.MODE_PRIVATE)
-            .edit()
-            .clear()
-            .apply()
         CountersStorageEngine.clearAllStores()
     }
